Improve event types by moving them to their own submodule, making them singletons to prevent needless equal instances, and by allowing extensions to expose their own event types.
