+++ betty/tests/model/test_ancestry.py
@@ -12,10 +12,8 @@ from betty.media_type import MediaType
 from betty.model import Entity
 from betty.model.ancestry import Person, Event, Place, File, Note, Presence, PlaceName, PersonName, Subject, Birth, \
     Enclosure, Described, Dated, HasPrivacy, HasMediaType, Link, HasLinks, HasNotes, HasFiles, Source, Citation, \
-    HasCitations, PresenceRole, Attendee, Beneficiary, Witness, EventType, UnknownEventType, LifeEventType, \
-    PreBirthEventType, PostDeathEventType, Baptism, Adoption, Death, Funeral, FinalDispositionEventType, Cremation, \
-    Burial, Will, Engagement, Marriage, MarriageAnnouncement, Divorce, DivorceAnnouncement, Residence, Immigration, \
-    Emigration, Correspondence, Occupation, Retirement, Confirmation, Missing, EVENT_TYPE_TYPES
+    HasCitations, PresenceRole, Attendee, Beneficiary, Witness, EventType
+from betty.model.event_type import Burial
 from betty.tests import TestCase
 
 
@@ -489,340 +487,6 @@ class PresenceTest(TestCase):
         event = Mock(Event)
         sut = Presence(Mock(Person), Mock(PresenceRole), event)
         self.assertEquals(event, sut.event)
-
-
-class EventTypeTest(TestCase):
-    def test_comes_before(self) -> None:
-        self.assertIsInstance(EventType.comes_before(), set)
-
-    def test_comes_after(self) -> None:
-        self.assertIsInstance(EventType.comes_after(), set)
-
-
-class UnknownEventTypeTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(UnknownEventType.name(), str)
-        self.assertNotEqual('', UnknownEventType.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = UnknownEventType()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-    def test_comes_before(self) -> None:
-        self.assertIsInstance(UnknownEventType.comes_before(), set)
-
-    def test_comes_after(self) -> None:
-        self.assertIsInstance(UnknownEventType.comes_after(), set)
-
-
-class PreBirthEventTypeTestCase(TestCase):
-    def test_comes_before(self) -> None:
-        self.assertIsInstance(PreBirthEventType.comes_before(), set)
-
-
-class LifeEventTypeTestCase(TestCase):
-    def test_comes_before(self) -> None:
-        self.assertIsInstance(LifeEventType.comes_before(), set)
-
-    def test_comes_after(self) -> None:
-        self.assertIsInstance(LifeEventType.comes_after(), set)
-
-
-class PostDeathEventTypeTestCase(TestCase):
-    def test_comes_after(self) -> None:
-        self.assertIsInstance(PostDeathEventType.comes_after(), set)
-
-
-class BirthTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Birth.name(), str)
-        self.assertNotEqual('', Birth.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Birth()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-    def test_comes_before(self) -> None:
-        self.assertIsInstance(Birth.comes_before(), set)
-
-
-class BaptismTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Baptism.name(), str)
-        self.assertNotEqual('', Baptism.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Baptism()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-
-class AdoptionTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Adoption.name(), str)
-        self.assertNotEqual('', Adoption.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Adoption()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-
-class DeathTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Death.name(), str)
-        self.assertNotEqual('', Death.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Death()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-    def test_comes_after(self) -> None:
-        self.assertIsInstance(Death.comes_after(), set)
-
-
-class FuneralTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Funeral.name(), str)
-        self.assertNotEqual('', Funeral.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Funeral()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-    def test_comes_after(self) -> None:
-        self.assertIsInstance(Funeral.comes_after(), set)
-
-
-class FinalDispositionEventTypeTestCase(TestCase):
-    def test_comes_after(self) -> None:
-        self.assertIsInstance(FinalDispositionEventType.comes_after(), set)
-
-
-class CremationTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Cremation.name(), str)
-        self.assertNotEqual('', Cremation.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Cremation()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-
-class BurialTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Burial.name(), str)
-        self.assertNotEqual('', Burial.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Burial()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-
-class WillTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Will.name(), str)
-        self.assertNotEqual('', Will.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Will()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-    def test_comes_after(self) -> None:
-        self.assertIsInstance(Will.comes_after(), set)
-
-
-class EngagementTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Engagement.name(), str)
-        self.assertNotEqual('', Engagement.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Engagement()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-    def test_comes_before(self) -> None:
-        self.assertIsInstance(Engagement.comes_before(), set)
-
-
-class MarriageTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Marriage.name(), str)
-        self.assertNotEqual('', Marriage.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Marriage()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-
-class MarriageAnnouncementTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(MarriageAnnouncement.name(), str)
-        self.assertNotEqual('', MarriageAnnouncement.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = MarriageAnnouncement()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-    def test_comes_before(self) -> None:
-        self.assertIsInstance(MarriageAnnouncement.comes_before(), set)
-
-
-class DivorceTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Divorce.name(), str)
-        self.assertNotEqual('', Divorce.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Divorce()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-    def test_comes_after(self) -> None:
-        self.assertIsInstance(Divorce.comes_after(), set)
-
-
-class DivorceAnnouncementTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(DivorceAnnouncement.name(), str)
-        self.assertNotEqual('', DivorceAnnouncement.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = DivorceAnnouncement()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-    def test_comes_before(self) -> None:
-        self.assertIsInstance(DivorceAnnouncement.comes_before(), set)
-
-    def test_comes_after(self) -> None:
-        self.assertIsInstance(DivorceAnnouncement.comes_after(), set)
-
-
-class ResidenceTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Residence.name(), str)
-        self.assertNotEqual('', Residence.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Residence()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-
-class ImmigrationTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Immigration.name(), str)
-        self.assertNotEqual('', Immigration.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Immigration()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-
-class EmigrationTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Emigration.name(), str)
-        self.assertNotEqual('', Emigration.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Emigration()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-
-class OccupationTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Occupation.name(), str)
-        self.assertNotEqual('', Occupation.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Occupation()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-
-class RetirementTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Retirement.name(), str)
-        self.assertNotEqual('', Retirement.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Retirement()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-
-class CorrespondenceTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Correspondence.name(), str)
-        self.assertNotEqual('', Correspondence.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Correspondence()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-
-class ConfirmationTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Confirmation.name(), str)
-        self.assertNotEqual('', Confirmation.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Confirmation()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-
-class MissingTestCase(TestCase):
-    def test_name(self) -> None:
-        self.assertIsInstance(Missing.name(), str)
-        self.assertNotEqual('', Missing.name())
-
-    def test_label(self) -> None:
-        with Translations(NullTranslations()):
-            sut = Missing()
-            self.assertIsInstance(sut.label, str)
-            self.assertNotEqual('', sut.label)
-
-
-class EventTypeTypesTest(TestCase):
-    def test(self) -> None:
-        for event_type_type in EVENT_TYPE_TYPES:
-            self.assertTrue(issubclass(event_type_type, EventType))
 
 
 class EventTest(TestCase):
